refactor(graph): unify transformer partition planning
