+++ frontend/Python/graph/__init__.py
@@ -28,15 +28,18 @@ from .partitioned_graph_driver import (
 from .partitioned_graph_driver import (
     SplitStrategy as SplitStrategy,
 )
-from .template_materialization import (
-    TemplateMaterializationPlan as TemplateMaterializationPlan,
-)
-from .template_materialization import TemplateUnit as TemplateUnit
-from .template_materialization import (
-    build_template_materialization_plan as build_template_materialization_plan,
-)
-from .template_partitioned_graph_driver import (
+from .partitioned_graph_driver import (
     TemplatePartitionedGraphDriver as TemplatePartitionedGraphDriver,
+)
+from .transformer_partition import (
+    TemplateInstanceBinding as TemplateInstanceBinding,
+)
+from .transformer_partition import TemplateUnit as TemplateUnit
+from .transformer_partition import (
+    TransformerPartitionPlan as TransformerPartitionPlan,
+)
+from .transformer_partition import (
+    build_transformer_partition_plan as build_transformer_partition_plan,
 )
 from .type import (
     DeviceType as DeviceType,
